refactor: 작은 수정
