+++ src/main/java/com/example/book_your_seat/payment/PaymentConst.java
@@ -3,7 +3,7 @@ package com.example.book_your_seat.payment;
 public final class PaymentConst {
     public PaymentConst() {}
 
-    public static final String ENTER_PAYMENTKEY = "PaymentKey 입력이 필요합니다.";
+    public static final String ENTER_PAYMENT_KEY = "PaymentKey 입력이 필요합니다.";
     public static final String ENTER_ORDER_ID = "orderId 입력이 필요합니다.";
     public static final String ENTER_AMOUNT = "amount 입력이 필요합니다.";
     public static final String ENTER_SEAT_ID = "seatId 입력이 필요합니다.";
