+++ src/main/java/com/example/book_your_seat/payment/service/facade/PaymentFacade.java
@@ -17,7 +17,7 @@ import com.example.book_your_seat.payment.service.dto.PaymentCommand;
 import com.example.book_your_seat.queue.service.facade.QueueService;
 import com.example.book_your_seat.reservation.domain.Reservation;
 import com.example.book_your_seat.reservation.domain.ReservationStatus;
-import com.example.book_your_seat.reservation.service.ReservationCommandService;
+import com.example.book_your_seat.reservation.service.command.ReservationCommandService;
 import com.example.book_your_seat.seat.service.query.SeatQueryService;
 import com.example.book_your_seat.user.domain.Address;
 import com.example.book_your_seat.user.domain.User;
