+++ src/main/java/com/example/book_your_seat/reservation/service/facade/ReservationFacade.java
@@ -2,7 +2,7 @@ package com.example.book_your_seat.reservation.service.facade;
 
 
 import com.example.book_your_seat.reservation.domain.Reservation;
-import com.example.book_your_seat.reservation.service.ReservationCommandService;
+import com.example.book_your_seat.reservation.service.command.ReservationCommandService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
