+++ src/main/java/com/example/book_your_seat/reservation/service/query/ReservationQueryService.java
@@ -1,4 +1,4 @@
-package com.example.book_your_seat.reservation.service;
+package com.example.book_your_seat.reservation.service.query;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
